+++ apps/test-app/app/routes/tests/text-box/index.spec.ts-snapshot.-visual-1-webkit-linux.png
