+++ apps/test-app/app/routes/tests/text-box/index.spec.ts-snapshot.-visual-composition-outline-1-chromium-linux.png
