+++ apps/test-app/app/routes/tests/text-box/index.spec.ts-snapshot.-visual-default-1-chromium-linux.png
