+++ apps/test-app/app/routes/tests/text-box/index.spec.ts-snapshot.-visual-default-1-webkit-linux.png
