+++ packages/kiwi-react/src/bricks/index.ts
@@ -18,6 +18,6 @@ export { Radio } from "./Radio.js";
 export { Switch } from "./Switch.js";
 export * as Tabs from "./Tabs.js";
 export { Textarea } from "./Textarea.js";
-export * as TextInput from "./TextInput.js";
+export * as TextInput from "./TextBox.js";
 export { Tooltip } from "./Tooltip.js";
 export { VisuallyHidden } from "./VisuallyHidden.js";
